571: v 2.3: add methods db
