+++ app/db/methods/edit_user.py
@@ -1,0 +1,8 @@
+from app.db.db_methods import users_collection
+
+
+def edit_user(user):
+    if users_collection.find_one_and_replace({'username': user.username}, user.pack()) is not None:
+        return True
+    else:
+        return False
