+++ app/routes/capacity.py
@@ -1,6 +1,6 @@
 from flask import Blueprint, request, current_app
 
-from app.db import db_methods
+from app.db.methods import file as file_methods
 
 
 capacity = Blueprint('capacity', __name__, template_folder='templates', static_folder='static')
@@ -10,7 +10,7 @@ capacity = Blueprint('capacity', __name__, template_folder='templates', static_f
 def system_capacity():
     units = {'b': 1, 'mb': 1024 ** 2, 'gb': 1024 ** 3}
     unit = units.get(request.args.get('unit', 'gb').lower(), units['gb'])
-    current_size = db_methods.get_storage()
+    current_size = file_methods.get_storage()
     ratio = current_size / current_app.config['MAX_SYSTEM_STORAGE']
     return {
         'size': current_size / unit,
