+++ app/routes/check_list.py
@@ -2,9 +2,12 @@ from flask import Blueprint, render_template, jsonify, request
 
 from app.utils import format_check_for_table, checklist_filter
 from app.db import db_methods
+from app.db.methods import check as check_methods
+
 from flask_login import login_required, current_user
 
 check_list = Blueprint('check_list', __name__, template_folder='templates', static_folder='static')
+
 
 @check_list.route("/")
 @login_required
@@ -35,10 +38,10 @@ def check_list_data():
     query = dict(filter=filter_query, limit=limit, offset=offset, sort=sort, order=order)
 
     if data.get("latest"):
-        rows, count = db_methods.get_latest_check_cursor(**query)
+        rows, count = check_methods.get_latest_check_cursor(**query)
     else:
         # get data and records count
-        rows, count = db_methods.get_checks_cursor(**query)
+        rows, count = check_methods.get_checks_cursor(**query)
 
     # construct response
     response = {
