+++ app/routes/criterion_packs.py
@@ -1,8 +1,7 @@
-
 from flask import Blueprint, render_template, abort
 from flask_login import login_required, current_user
 
-from app.db import db_methods
+from app.db.methods import criteria_pack as criteria_pack_methods
 
 
 criterion_packs = Blueprint('criterion_packs', __name__, template_folder='templates', static_folder='static')
@@ -13,6 +12,5 @@ criterion_packs = Blueprint('criterion_packs', __name__, template_folder='templa
 def criteria_packs():
     if not current_user.is_admin:
         abort(403)
-    packs = db_methods.get_criterion_pack_list()
+    packs = criteria_pack_methods.get_criterion_pack_list()
     return render_template('./pack_list.html', packs=packs, name=current_user.name, navi_upload=True)
-
