+++ app/routes/get_zip.py
@@ -10,10 +10,14 @@ from flask import Blueprint, abort, request, Response
 from flask_login import login_required, current_user
 
 from app.db import db_methods
+from app.db.methods import file as file_methods
+from app.db.methods import check as check_methods
+
 from app.utils import checklist_filter, format_check_for_table
 from app.server_consts import URL_DOMEN
 
 get_zip = Blueprint('get_zip', __name__, template_folder='templates', static_folder='static')
+
 
 def get_query(req):
     # query for download csv/zip
@@ -30,7 +34,7 @@ def get_query(req):
 
 
 def get_stats():
-    rows, count = db_methods.get_checks(**get_query(request))
+    rows, count = check_methods.get_checks(**get_query(request))
     return [format_check_for_table(item, set_link=URL_DOMEN) for item in rows]
 
 
@@ -46,10 +50,10 @@ def get_zip_main():
     dirpath = tempfile.TemporaryDirectory()
 
     # write files
-    checks_list, _ = db_methods.get_checks(**get_query(request))
+    checks_list, _ = check_methods.get_checks(**get_query(request))
     for check in checks_list:
-        db_file = db_methods.find_pdf_by_file_id(check['_id'])
-        original_name = db_methods.get_check(check['_id']).filename #get a filename from every check
+        db_file = file_methods.find_pdf_by_file_id(check['_id'])
+        original_name = check_methods.get_check(check['_id']).filename #get a filename from every check
         if db_file is not None:
             final_name = original_name if (original_name and original_names) else db_file.filename
             # to avoid overwriting files with one name and different content: now we save only last version of pres (from last check)
