+++ app/routes/results.py
@@ -3,7 +3,9 @@ from bson import ObjectId
 
 from flask import Blueprint, render_template
 
-from app.db import db_methods
+from app.db.methods import check as check_methods
+from app.db.methods import celery_check as celery_check_methods
+
 from app.utils import format_check
 from app.root_logger import get_root_logger
 
@@ -20,10 +22,10 @@ def results_main(_id):
     except bson.errors.InvalidId:
         logger.error('_id exception:', exc_info=True)
         return render_template("./404.html")
-    check = db_methods.get_check(oid)
+    check = check_methods.get_check(oid)
     if check is not None:
         # show processing time for user
-        avg_process_time = None if check.is_ended else db_methods.get_average_processing_time()
+        avg_process_time = None if check.is_ended else celery_check_methods.get_average_processing_time()
         return render_template("./results.html", navi_upload=True, results=check,
                                columns=TABLE_COLUMNS, avg_process_time=avg_process_time,
                                stats=format_check(check.pack()))
