+++ app/routes/tasks.py
@@ -7,7 +7,9 @@ from flask_login import login_required, current_user
 from app.root_logger import get_root_logger
 
 from app.utils import get_file_len, check_file
-from app.db import db_methods
+from app.db.methods import file as file_methods
+from app.db.methods import check as check_methods
+from app.db.methods import celery_check as celery_check_methods
 from app.db.types.Check import Check
 
 from app.server_consts import ALLOWED_EXTENSIONS, UPLOAD_FOLDER
@@ -29,7 +31,7 @@ def run_task():
     if not file:
         logger.critical("request doesn't include file")
         return "request doesn't include file"
-    if get_file_len(file) * 2 + db_methods.get_storage() > current_app.config['MAX_SYSTEM_STORAGE']:
+    if get_file_len(file) * 2 + file_methods.get_storage() > current_app.config['MAX_SYSTEM_STORAGE']:
         logger.critical('Storage overload has occured')
         return 'storage_overload'
     file_check_response = check_file(file, extension, ALLOWED_EXTENSIONS[file_ext_type], check_mime=True)
@@ -52,10 +54,10 @@ def run_task():
     filepath = join(UPLOAD_FOLDER, f"{file_id}.{extension}")
     file.save(filepath)
     # add file and file's info to db
-    file_id = db_methods.add_file_info_and_content(current_user.username, filepath, file_type, file_id)
+    file_id = file_methods.add_file_info_and_content(current_user.username, filepath, file_type, file_id)
     # use pdf from response or convert to pdf and save on disk and db
     if current_user.two_files and pdf_file:
-        if get_file_len(pdf_file) * 2 + db_methods.get_storage() > current_app.config['MAX_SYSTEM_STORAGE']:
+        if get_file_len(pdf_file) * 2 + file_methods.get_storage() > current_app.config['MAX_SYSTEM_STORAGE']:
             logger.critical('Storage overload has occured')
             return 'storage_overload'
         logger.info(
@@ -63,9 +65,9 @@ def run_task():
         filenamepdf, extension = pdf_file.filename.rsplit('.', 1)
         filepathpdf = join(UPLOAD_FOLDER, f"{file_id}.{extension}")
         pdf_file.save(filepathpdf)
-        converted_id = db_methods.add_file_to_db(filenamepdf, filepathpdf)
+        converted_id = file_methods.add_file_to_db(filenamepdf, filepathpdf)
     else:
-        converted_id = db_methods.write_pdf(filename, filepath)
+        converted_id = file_methods.write_pdf(filename, filepath)
     check = Check({
         '_id': file_id,
         'conv_pdf_fs_id': converted_id,
@@ -80,10 +82,11 @@ def run_task():
         'is_failed': False,
         'params_for_passback': current_user.params_for_passback
     })
-    db_methods.add_check(file_id, check)  # add check for parsed_file to db
+    check_methods.add_check(file_id, check)  # add check for parsed_file to db
     task = create_task.delay(check.pack(to_str=True))  # add check to queue
-    db_methods.add_celery_task(task.id, file_id)  # mapping celery_task to check (check_id = file_id)
+    celery_check_methods.add_celery_task(task.id, file_id)  # mapping celery_task to check (check_id = file_id)
     return {'task_id': task.id, 'check_id': str(file_id)}
+
 
 @tasks.route("/<task_id>", methods=["GET"])
 @login_required
